Update web-animation.json
Regarding the changes to Chrome status (which I expect are the only controversial parts of this change), the linked to page refers to playback control. There is much more to the API than just playback control. Furthermore, that part of the spec has changed recently so whatever is being shipped is wrong.