Changing dpcm back to yes again
dpcm is supported, as per a previous pull request I made. The feature test fails as there is a bug (listed here) where values <1 don't fire. But in reality, as screens with less than 1dpcm don't exist, this isn't a bug you'd hit in the wild except for feature detects like this.